Make audio address handling 64-bit-safe (pointer hygiene, stage 3d)
Two kinds of pointer casts in AudioDaemon, both broken by a 64-bit build:

- Real fixes: two sanity checks cast a CSample* to uint32 and test it
  against a small threshold ((uint32)psam > 32 / > 3) to tell a genuine
  pointer apart from the psamINSTANCE_* sentinels (CSample*)0x1..0x3. On
  x64 a real pointer whose low 32 bits happen to be <= 3 would be
  misread as a sentinel, so these must compare the full pointer. Widened
  to uintptr.

- Lossy identity ids: u4_id (built from CInstance* pins) and u4SoundID =
  (uint32)psam pack a pointer into a uint32 used only as an opaque key
  (never turned back into a pointer - u4SoundID just flows into u4AudioID
  and equality checks). Keep the 32-bit id design but make the narrowing
  explicit and warning-free with (uint32)(uintptr)ptr. The uint32/handle
  operands beside them (smatSound = TSoundMaterial, sndhndSample =
  TSoundHandle, both uint32) are left as-is - they are not pointers.

AudioADPCM.cpp needs nothing: its casts are all sample-value conversions
or sizeof pointer bumps, and its one (uint32)(pu1_dst - pu1_dst_start) is
a decoded-length narrowing that already fits.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Fourth slice of the uint32->uintptr
sweep after VirtualMem (ba41823), FastHeap (3825cd3), loaders (43c4b87).

+++ jp2_pc/Source/Lib/Audio/AudioDaemon.cpp
@@ -2023,7 +2023,7 @@ void CAudioDaemon::Process(const CMessageCollision& msg)
 	if ( /*(msg.fEnergyMax == 0.0) &&*/ (msg.fEnergySlide > 0.0) && (pcol->bSlide()) )
 	{
 		// Make a quick ID that should be unique for all sliding collisions.
-		uint32 u4_id = (uint32)msg.smatSound1 + (uint32)msg.smatSound2 + (uint32)msg.pins1 + (uint32)msg.pins2;
+		uint32 u4_id = (uint32)msg.smatSound1 + (uint32)msg.smatSound2 + (uint32)(uintptr)msg.pins1 + (uint32)(uintptr)msg.pins2;
 
 /*		dprintf("Frame: %d\n\n", CMessageStep::u4Frame);
 		dprintf("Slide Mat 1: %x\n", (uint32)msg.smatSound1);
@@ -2142,7 +2142,7 @@ void CAudioDaemon::Process(const CMessageMove& msg_move)
 				CVector3<>		v3_up		= d3ZAxis * p3_pres.r3Rot;	// up vector
 
 #if VER_TEST
-				AlwaysAssert((uint32)asatSoundObjects[u4_sample_count].psam > 32);
+				AlwaysAssert((uintptr)asatSoundObjects[u4_sample_count].psam > 32);
 #endif
 
 				asatSoundObjects[u4_sample_count].psam->SetPosition(p3_pres.v3Pos.tX,p3_pres.v3Pos.tZ,p3_pres.v3Pos.tY, false);
@@ -2914,7 +2914,7 @@ bool CAudioDaemon::bPlaySoundEffect
 		return true;
 
 	// make an Id that identifies this instance of this sample
-	uint32 u4_id = (uint32)msg.sndhndSample + (uint32)msg.pinsParent;
+	uint32 u4_id = (uint32)msg.sndhndSample + (uint32)(uintptr)msg.pinsParent;
 
 	// If this is a looped effect then chekc the collision loop flags
 	if (msg.bLooped) 
@@ -2944,7 +2944,7 @@ bool CAudioDaemon::bPlaySoundEffect
 
 	/// the cahce entry we may have found could have the wrong buffer type..
 #if VER_DEBUG
-	if (((uint32)psam>3) && CAudio::pcaAudio->bAudioActive())
+	if (((uintptr)psam>3) && CAudio::pcaAudio->bAudioActive())
 	{
 		switch (msg.u4SpatialType)
 		{
@@ -3009,7 +3009,7 @@ bool CAudioDaemon::bPlaySoundEffect
 
 		if (pser)
 		{
-			pser->u4SoundID = (uint32)psam;
+			pser->u4SoundID = (uint32)(uintptr)psam;
 		}
 
 		// if we are playing a dino effect then we need to mute any ambient sounds
